Update deploy scripts
Hard coded in the current defacto dev instance. We can update the build
file again when the normal crypto-proxy-xx hosts are set up there. I
also updated the staging host instances (of which there is one). The
copyTarballToDistributions() method was screwing up the deploy by
truncating the actual deployment archive, so I removed it. It seems
applicable only to a different build process.